Refactor HistoricalStorage into cohesive submodules
Extract configuration and statistics from `src/storage/historical.rs` into
dedicated submodules `config.rs` and `stats.rs` to improve maintainability
and reduce module bloat. The public API remains compatible via re-exports.

- Created `src/storage/historical/` module directory
- Extracted `RetentionPolicy`, `PreAnchorHook` to `config.rs`
- Extracted `HistoricalStats`, `CacheMetrics` to `stats.rs`
- Cleaned up imports and visibility in `mod.rs`
