+++ src/storage/historical/mod.rs
@@ -27,130 +27,16 @@ use std::sync::atomic::{AtomicU64, Ordering};
 #[cfg(feature = "observability")]
 use tracing;
 
-/// Default maximum number of versions per entity (DoS protection)
-pub const DEFAULT_MAX_VERSIONS_PER_ENTITY: usize = 1_000;
-
-/// Default maximum age for versions in milliseconds (365 days)
-pub const DEFAULT_MAX_VERSION_AGE_MS: i64 = 365 * 24 * 60 * 60 * 1000;
-
-/// Maximum recursion depth for version reconstruction (DoS protection).
-///
-/// This limit prevents stack overflow from corrupted version chains or cycles.
-/// A depth of 100 is sufficient for any legitimate use case since anchors are
-/// typically created every 10-20 versions.
-pub const MAX_RECONSTRUCTION_DEPTH: usize = 100;
-
-/// Retention policy for version history (DoS protection).
-///
-/// Controls how many versions are kept and for how long to prevent
-/// unbounded memory growth from malicious or buggy clients.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct RetentionPolicy {
-    /// Maximum number of versions to keep per entity
-    pub max_versions_per_entity: usize,
-    /// Maximum age of versions in milliseconds (older versions are pruned)
-    pub max_age_ms: i64,
-}
-
-impl Default for RetentionPolicy {
-    fn default() -> Self {
-        RetentionPolicy {
-            max_versions_per_entity: DEFAULT_MAX_VERSIONS_PER_ENTITY,
-            max_age_ms: DEFAULT_MAX_VERSION_AGE_MS,
-        }
-    }
-}
-
-impl RetentionPolicy {
-    /// Create a new retention policy with custom limits
-    pub fn new(max_versions_per_entity: usize, max_age_ms: i64) -> Self {
-        RetentionPolicy {
-            max_versions_per_entity,
-            max_age_ms,
-        }
-    }
-
-    /// Create a policy with no retention limits (unbounded)
-    pub fn unbounded() -> Self {
-        RetentionPolicy {
-            max_versions_per_entity: usize::MAX,
-            max_age_ms: i64::MAX,
-        }
-    }
-}
-
-/// Pre-anchor hook for creating snapshots before anchor storage.
-///
-/// This hook is called **before** storing an anchor to create synchronized snapshots
-/// and return a snapshot ID to be stored atomically with the anchor. This enables
-/// strong consistency for provenance tracking between graph versioning and vector indexing.
-///
-/// # Arguments
-///
-/// * `entity_type` - Type of entity ("node" or "edge")
-/// * `entity_id` - ID of the entity as u64
-/// * `timestamp` - Transaction timestamp when the anchor is being created
-/// * `properties` - Property map that will be stored in the anchor
-///
-/// # Returns
-///
-/// * `Ok(Some(snapshot_id))` - Snapshot created successfully, link to anchor
-/// * `Ok(None)` - No snapshot needed or empty index
-/// * `Err(e)` - Snapshot creation failed (anchor will still be created, graceful degradation)
-///
-/// # Examples
-///
-/// ```ignore
-/// use gallifreydb::storage::historical::PreAnchorHook;
-/// use std::sync::Arc;
-///
-/// let hook: PreAnchorHook = Arc::new(|entity_type, entity_id, timestamp, properties| {
-///     // Create vector snapshot before anchor is stored
-///     let snapshot_id = create_vector_snapshot(timestamp, properties)?;
-///     Ok(Some(snapshot_id))
-/// });
-/// ```
-pub type PreAnchorHook = Arc<
-    dyn Fn(
-            /* entity_type */ &str,
-            /* entity_id */ u64,
-            /* timestamp */ Timestamp,
-            /* properties */ &PropertyMap,
-        ) -> Result<Option<usize>>
-        + Send
-        + Sync,
->;
-
-/// Context for pre-anchor hook invocation.
-///
-/// Groups the parameters needed to invoke a pre-anchor hook, improving
-/// readability and reducing the number of function parameters.
-struct AnchorHookContext<'a> {
-    /// Entity type ("node" or "edge")
-    entity_type: &'a str,
-    /// Entity ID as u64
-    entity_id: u64,
-    /// Transaction timestamp
-    timestamp: Timestamp,
-    /// Properties being stored in the anchor
-    properties: &'a PropertyMap,
-}
-
-/// Default cache size for reconstructed properties (10,000 entries)
-const DEFAULT_RECONSTRUCTION_CACHE_SIZE: usize = 10_000;
-
-/// Anchor cache size ratio relative to main cache (Improvement #1: Issue #338).
-///
-/// Typically 10-20% of versions become anchors depending on `anchor_interval`.
-/// With default interval of 10, we get ~10% anchors. Setting to 1/5 (20%)
-/// provides headroom for configurations with smaller intervals.
-const ANCHOR_CACHE_SIZE_RATIO: usize = 5; // 20% of main cache
-
-/// Minimum anchor cache size to ensure reasonable performance (Improvement #1: Issue #338).
-///
-/// Even with very small main caches, we want enough anchor cache to hold
-/// at least a few anchors to avoid immediate evictions.
-const MIN_ANCHOR_CACHE_SIZE: usize = 100;
+/// Configuration types and constants for historical storage.
+pub mod config;
+pub use config::{
+    DEFAULT_MAX_VERSIONS_PER_ENTITY, DEFAULT_MAX_VERSION_AGE_MS, MAX_RECONSTRUCTION_DEPTH,
+    PreAnchorHook, RetentionPolicy,
+};
+use config::{
+    ANCHOR_CACHE_SIZE_RATIO, AnchorHookContext, DEFAULT_RECONSTRUCTION_CACHE_SIZE,
+    MIN_ANCHOR_CACHE_SIZE,
+};
 
 /// Historical storage for versioned nodes and edges.
 ///
@@ -2380,168 +2266,8 @@ impl Default for HistoricalStorage {
     }
 }
 
-/// Cache performance metrics (Issue #338: Improvement #3).
-///
-/// Provides granular insight into cache behavior:
-/// - `primary_cache_hits`: Fast path hits (most common)
-/// - `anchor_cache_hits`: Fallback hits (indicates primary cache pressure)
-/// - `full_reconstructions`: Slow path (indicates insufficient cache capacity)
-///
-/// # Interpretation
-/// - High `anchor_cache_hits` + low `primary_cache_hits` → increase primary cache size
-/// - High `full_reconstructions` → increase overall cache capacity
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct CacheMetrics {
-    /// Number of successful lookups in primary property cache (fast path)
-    pub primary_cache_hits: u64,
-    /// Number of successful lookups in anchor cache fallback
-    pub anchor_cache_hits: u64,
-    /// Number of full property reconstructions from deltas
-    pub full_reconstructions: u64,
-}
-
-impl CacheMetrics {
-    /// Calculate total cache operations (hits + reconstructions).
-    pub fn total_operations(&self) -> u64 {
-        self.primary_cache_hits + self.anchor_cache_hits + self.full_reconstructions
-    }
-
-    /// Calculate overall cache hit rate (0.0 to 1.0).
-    ///
-    /// Returns None if no operations have been performed yet.
-    pub fn hit_rate(&self) -> Option<f64> {
-        let total = self.total_operations();
-        if total == 0 {
-            None
-        } else {
-            Some((self.primary_cache_hits + self.anchor_cache_hits) as f64 / total as f64)
-        }
-    }
-
-    /// Calculate primary cache hit rate (0.0 to 1.0).
-    ///
-    /// This shows how often the primary cache is sufficient without fallback.
-    /// Returns None if no operations have been performed yet.
-    pub fn primary_hit_rate(&self) -> Option<f64> {
-        let total = self.total_operations();
-        if total == 0 {
-            None
-        } else {
-            Some(self.primary_cache_hits as f64 / total as f64)
-        }
-    }
-
-    /// Calculate anchor cache fallback rate (0.0 to 1.0).
-    ///
-    /// This shows how often we need to fall back to the anchor cache.
-    /// High values indicate the primary cache is under pressure.
-    pub fn anchor_fallback_rate(&self) -> Option<f64> {
-        let total = self.total_operations();
-        if total == 0 {
-            None
-        } else {
-            Some(self.anchor_cache_hits as f64 / total as f64)
-        }
-    }
-
-    /// Calculate reconstruction rate (0.0 to 1.0).
-    ///
-    /// This shows how often we need to perform full reconstruction.
-    /// High values indicate insufficient overall cache capacity.
-    pub fn reconstruction_rate(&self) -> Option<f64> {
-        let total = self.total_operations();
-        if total == 0 {
-            None
-        } else {
-            Some(self.full_reconstructions as f64 / total as f64)
-        }
-    }
-}
-
-/// Statistics about the historical storage.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct HistoricalStats {
-    /// Total number of node versions stored
-    pub total_node_versions: usize,
-    /// Total number of edge versions stored
-    pub total_edge_versions: usize,
-    /// Number of anchor node versions
-    pub node_anchor_count: usize,
-    /// Number of delta node versions
-    pub node_delta_count: usize,
-    /// Number of anchor edge versions
-    pub edge_anchor_count: usize,
-    /// Number of delta edge versions
-    pub edge_delta_count: usize,
-    /// Number of unique nodes with version history
-    pub unique_nodes: usize,
-    /// Number of unique edges with version history
-    pub unique_edges: usize,
-    /// Number of cached node property reconstructions (regular cache)
-    pub node_cache_entries: usize,
-    /// Number of cached edge property reconstructions (regular cache)
-    pub edge_cache_entries: usize,
-    /// Number of cached node anchor properties (dedicated anchor cache, Issue #338)
-    pub node_anchor_cache_entries: usize,
-    /// Number of cached edge anchor properties (dedicated anchor cache, Issue #338)
-    pub edge_anchor_cache_entries: usize,
-}
-
-impl HistoricalStats {
-    /// Calculate the compression ratio (anchors vs total versions).
-    pub fn compression_ratio(&self) -> f64 {
-        let total_versions = self.total_node_versions + self.total_edge_versions;
-        let total_anchors = self.node_anchor_count + self.edge_anchor_count;
-
-        if total_versions == 0 {
-            return 1.0;
-        }
-
-        total_anchors as f64 / total_versions as f64
-    }
-
-    /// Estimate total cache memory usage in bytes (Issue #338: Memory Accounting).
-    ///
-    /// Provides rough estimate of memory consumed by all caches. Actual memory
-    /// usage may vary based on property sizes, Arc overhead, and allocator behavior.
-    ///
-    /// # Formula
-    /// Per entry overhead:
-    /// - VersionId: 8 bytes
-    /// - Arc pointer: 8 bytes
-    /// - PropertyMap overhead: ~16 bytes
-    /// - Average property data: ~100 bytes (varies by use case)
-    /// - Total: ~132 bytes per entry (rounded to 150 for safety margin)
-    ///
-    /// # Returns
-    /// Estimated bytes used by all caches (primary + anchor)
-    ///
-    /// # Example
-    /// ```no_run
-    /// # use gallifreydb::storage::historical::HistoricalStorage;
-    /// let storage = HistoricalStorage::new();
-    /// // ... perform operations ...
-    /// let stats = storage.stats();
-    /// let bytes = stats.estimated_cache_memory_bytes();
-    /// println!("Cache using ~{:.2} MB", bytes as f64 / 1_048_576.0);
-    /// ```
-    pub fn estimated_cache_memory_bytes(&self) -> usize {
-        // Rough estimate per cache entry:
-        // - VersionId (u64): 8 bytes
-        // - Arc<PropertyMap> pointer: 8 bytes
-        // - PropertyMap struct overhead: ~16 bytes
-        // - Average property data: ~100 bytes (varies widely)
-        // Total: ~132 bytes, rounded to 150 for safety margin
-        const BYTES_PER_ENTRY: usize = 150;
-
-        let total_entries = self.node_cache_entries
-            + self.edge_cache_entries
-            + self.node_anchor_cache_entries
-            + self.edge_anchor_cache_entries;
-
-        total_entries * BYTES_PER_ENTRY
-    }
-}
+pub mod stats;
+pub use stats::{CacheMetrics, HistoricalStats};
 
 #[cfg(test)]
 mod tests {
